backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
authlib==1.3.2
httpx==0.27.2
redis[asyncio]==5.1.1
python-multipart==0.0.12
sqlalchemy-utils==0.41.2
pytest==8.3.3
pytest-asyncio==0.24.0
factory-boy==3.3.1
apscheduler==3.10.4
